+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Unhandled exception in telemetry when username could not be inferred on Windows
+- Metadata is now correctly updated for hybrid spaces
 
 ## [0.7.3] - 2024-02-09
 ### Added